Removed all external files because all of the functions have been implemented in .cc files of Worker and TaskGenerator and also switched extension from Veins to VeinsInet
