+++ src/app/TaskGenerator.h
@@ -22,7 +22,7 @@
 
 #pragma once
 
-#include "veins/veins.h"
+#include "veins_inet/veins_inet.h"
 
 #include "app/Task.h"
 #include "app/messages/HelpMessage_m.h"
@@ -32,7 +32,11 @@
 #include "app/vehiclesHandling/HelperVehicleInfo.h"
 #include "app/loadBalancing/sortingAlgorithm/BaseSorting.h"
 #include "loadBalancing/BusState.h"
-#include "veins/modules/application/ieee80211p/DemoBaseApplLayer.h"
+#include "veins_inet/VeinsInetApplicationBase.h"
+#include "inet/transportlayer/udp/UdpHeader_m.h"
+#include "inet/networklayer/ipv4/Ipv4InterfaceData.h"
+#include "inet/networklayer/common/L3AddressResolver.h"
+#include "inet/networklayer/ipv4/Ipv4.h"
 
 using namespace omnetpp;
 
@@ -48,7 +52,7 @@ namespace task_offloading {
  *
  */
 
-class VEINS_API TaskGenerator : public veins::DemoBaseApplLayer {
+class VEINS_INET_API TaskGenerator : public veins::VeinsInetApplicationBase {
 public:
     void initialize(int stage) override;
     void finish() override;
@@ -85,16 +89,14 @@ protected:
     BaseSorting* loadBalancingAlgorithm;
 
 protected:
-    void onBSM(veins::DemoSafetyMessage* bsm) override;
-    void onWSM(veins::BaseFrame1609_4* wsm) override;
-    void onWSA(veins::DemoServiceAdvertisment* wsa) override;
-
-    void handleSelfMsg(cMessage* msg) override;
     void handleAvailabilityMessage(AvailabilityMessage* okMsg);
     void handleResponseMessage(ResponseMessage* responseMsg);
     void sendAgainData(DataMessage* data);
     void balanceLoad();
     void vehicleHandler();
-    void handlePositionUpdate(cObject* obj) override;
+    virtual void handleStartOperation(inet::LifecycleOperation* doneCallback) override;
+    virtual void handleStopOperation(inet::LifecycleOperation* doneCallback) override;
+    virtual void handleMessageWhenUp(inet::cMessage* msg) override;
+    virtual void sendUnicastPacket(std::unique_ptr<inet::Packet> pk, inet::L3Address destAddress, int portNumber);
 };
 }
